requirements.txt
python-arango==7.5.2
multithread-parallel-processing==1.0.0
numpy==1.26.4
click==8.0.1
notebook==7.2.0
psycopg-binary==3.1.18
pymongo==4.3.3
python-dotenv==1.0.0
seaborn==0.13.2
imblearn==0.0
scikit-learn==1.4.2
matplotlib==3.8.4
optuna==3.2.0
joblib==1.3.2
tqdm==4.66.5
texttable==1.7.0
web3==6.19.0
psycopg2-binary==2.9.9
python-cli-scheduler==1.1.1

ipython==8.18.1
pymupdf==1.24.5
pgvector==0.2.5
asyncpg==0.29.0
sqlalchemy[asyncio]==1.4.52
greenlet==3.0.3
diffusers==0.29.0
transformers==4.41.2
accelerate==0.31.0
peft==0.11.1
flasgger==0.9.7.1
flask_restful==0.3.10
flask_cors==4.0.1
langdetect==1.0.9
emoji==2.12.1
ruptures==1.1.9
adtk==0.6.2
requests==2.32.3

contractions==0.1.73
gensim==4.3.3
datasets==2.21.0
dspy-ai==2.4.14

qdrant-client==1.12.1
optimum[onnxruntime]

langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14

sanic==24.6.0
Sanic-Cors==2.2.0
sanic-ext==23.12.0
sanic-routing==23.12.0
sanic-testing==23.12.0

termcolor==2.3.0
tweepy==4.14.0
PyJWT==2.8.0
prometheus_client==0.17.1
elasticsearch==8.10.1
telethon==1.34.0
moviepy==1.0.3
redis==5.0.3

requests~=2.32.3
prometheus-client~=0.17.1
APScheduler~=3.11.0

spacy==3.8.2
keybert==0.8.5
nltk==3.2.4
